// cgocore-ui/src/app/features/models/bookings/views/basic/DatedFlightBookingBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightBookingBasicView {

id: number;
selected: boolean;
canDelete: boolean;
datedFlightCompanyIataCode: string;
datedFlightFlightNumber: string;
datedFlightDepartureDate: string;
bookingAirWaybillAirWaybillNumber: string;
numberOfPieces: number;
grossWeight: number;
grossWeightUnitOfMeasureCode: string;
grossVolume: number;
grossVolumeUnitOfMeasureCode: string;

/* Specific Code Start */
/* Specific Code End */
}