// cgocore-ui/src/app/features/models/bookings/filters/DatedFlightBookingFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightBookingFilter {

datedFlightCompanyIataCode: string;
datedFlightFlightNumber: string;
datedFlightDepartureDateMinValue: string;
datedFlightDepartureDateMaxValue: string;
bookingAirWaybillAirWaybillNumber: string;

/* Specific Code Start */
/* Specific Code End */
}