// cgocore-ui/src/app/features/models/bookings/sortings/DatedFlightBookingSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightBookingSorting {

datedFlightCompanyIataCodeOrderType: string;
datedFlightFlightNumberOrderType: string;
datedFlightDepartureDateOrderType: string;
bookingAirWaybillAirWaybillNumberOrderType: string;
numberOfPiecesOrderType: string;
grossWeightOrderType: string;
grossWeightUnitOfMeasureCodeOrderType: string;
grossVolumeOrderType: string;
grossVolumeUnitOfMeasureCodeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}